+++ .eslintrc.js
@@ -1,13 +1,13 @@
 module.exports = {
   parser: '@typescript-eslint/parser',
   parserOptions: {
-    ecmaVersion:  2018,  // Allows for the parsing of modern ECMAScript features
+    ecmaVersion:  2020,  // Allows for the parsing of modern ECMAScript features
     sourceType:  'module',
   },
   extends: [
     'plugin:@typescript-eslint/recommended',
-    'prettier',
     'prettier/@typescript-eslint',
+    'prettier'
   ],
   plugins: ['@typescript-eslint', 'prettier'],
   rules: {